cars_service/pyproject.toml
[tool.poetry]
name = "cars-service"
version = "0.1.0"
description = ""
authors = ["Illia Troshchynskyi <itroshchinskiy@rambler.ua>"]
readme = "README.md"
packages = [{include = "cars_service"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.101.1"}
python-dotenv = "^1.0.0"
alembic = "^1.10.4"
pydantic = "^2.2.0"
pydantic-settings = "^2.0.3"
python-multipart = "^0.0.6"
aiofiles = "^23.2.1"
psycopg = "^3.1.10"
sqlalchemy = "^2.0.20"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
flake8-print = "^5.0.0"
pytest = "^7.3.1"
pytest-asyncio = "^0.21.0"
pytest-httpx = "^0.22.0"
flake8-quotes = "^3.3.2"
coverage = "^7.3.0"
polyfactory = "^2.7.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
